+++ client/ui/Credits.jsx
@@ -30,8 +30,10 @@ export default class Credits extends Component {
                       <h5><b>Release information</b></h5>
 
                       <ul>
-                        <b>6.16: </b>
+                        <b>6.16.1: </b>
                         <li>Multiword search improved</li>
+                        <li>Urdu is now enabled in default settings</li>
+                        <li>Multiword query boosted</li>
                       </ul>
 
                       <ul>