+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+## 6.16.1:
+  * Urdu is now enabled in default settings
+  * Multiword query boosted
+
 ## 6.16:
   * Multiword search improved (boost removed)
+
 
 ## 6.15.2:
   * Search scope configuration now retains in URL making results shareable